Moved Files Into Directories
Closes #13
