+++ src/processing/extraction.test.js
