+++ src/routes/ViewPuzzle.css
